requirements.txt
streamlit
feedparser
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
chromadb
pypdf
beautifulsoup4
python-dotenv
requests
google-generativeai
